Refactor wiki service
